requirements.txt
eventlet==0.29.0
flask==1.1.2
flask-socketio==4.3.1
python-socketio==4.6.0
requests==2.24.0
